Extract MissingParameterWrapper wrapper from Templater
We use this from both the `Templater` and `MagicNamespace`, which now
live in different files, so it seems sensible for this to live in its
own file used by both rather than live in the `Templater` and just
happen to be found by the `MagicNamespace`.
